gitea/Python/assignments/assignment-05-final-project/template/src/main.py
#!/usr/bin/env python3
"""
项目主入口
TODO: 实现你的项目功能
"""

import argparse
import sys
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()


def main():
    """主函数"""
    parser = argparse.ArgumentParser(
        description="你的项目描述",
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog="""
示例:
  python src/main.py --help          显示帮助
  python src/main.py [命令] [参数]    执行功能
        """
    )
    
    # TODO: 添加你的命令行参数
    # parser.add_argument("command", help="命令")
    # parser.add_argument("--option", help="选项")
    
    args = parser.parse_args()
    
    # TODO: 实现你的主逻辑
    print("🚧 项目待实现")
    print("请修改 src/main.py 实现你的功能")


if __name__ == "__main__":
    main()

